third_party/WebKit/LayoutTests/http/tests/devtools/elements/modify-chardata-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that elements panel updates dom tree structure upon changes to characters.


Running: testDumpInitial
========= Original for normal mutation ========
  <div id="node">Foo</div>

Running: testSetAttribute
===== Mutated text node =====
  <div id="node">Bar</div>

Running: testModifyViaRange
========= Original for Range mutation ========
  <div id="rangenode">Lorem ipsum dolor sit amet</div>
===== Contents modified via Range =====
- <div id="rangenode">
      "Lorem ips"
      <span>test</span>
      "lor sit amet"
  </div>
